27 July Improved ReadScannerInput.java

+++ ReadScannerInput/src/ReadScannerInput.java
@@ -5,14 +5,14 @@ public class ReadScannerInput {
         Scanner scanner = new Scanner(System.in);
 
         int counter = 1;
-        int sum = 0;
+        double sum = 0;
 
 
         while (counter <= 5) {
             System.out.println("Enter number #" + counter + ":");
             String nextNumber = scanner.nextLine();
             try {
-                int number = Integer.parseInt(nextNumber);
+                double number = Double.parseDouble(nextNumber);
                 counter++;
                 sum += number;
